pkg/openthread: cleanup of code

+++ pkg/openthread/contrib/platform_random.c
@@ -29,7 +29,7 @@ void ot_random_init(void)
     char cpu_id[CPUID_LEN];
     cpuid_get(cpu_id);
     uint32_t seed = 0;
-    for (int i = 0; i < (int) CPUID_LEN; i++) {
+    for (unsigned i = 0; i < CPUID_LEN; i++) {
         seed += cpu_id[i];
     }
     random_init(seed);
